first and final commit HTML and the CSS
Added the height for the html. This style rule is for the body content. the max and min width is for the browsing range. Made the margins auto to center the page content. The width is for the browsing range. The style rule is for the header, ul, nav, li and its the border box model.The class row and the pseudo element style rule is for to make sure that there is not a container collapse.The image that is in the list items thats in the comments id. The list items in the comments id and is displayed in block. changed the css files from txt to just .css and changed the home page file to index.html.